Break out Repo access from API.DocsController
Also:

* add HexWeb.Assets
* add HexWeb.Sitemaps
* add HexWeb.DocsTar
